refactor: migrate common files to droplet-api and fix formatting
