+++ controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/AuthClientRepository.kt
@@ -2,9 +2,10 @@ package app.simplecloud.controller.runtime.oauth
 
 import app.simplecloud.controller.runtime.Repository
 import app.simplecloud.controller.runtime.database.Database
-import app.simplecloud.controller.shared.auth.Scope
 import app.simplecloud.controller.shared.db.tables.records.Oauth2ClientDetailsRecord
 import app.simplecloud.controller.shared.db.tables.references.OAUTH2_CLIENT_DETAILS
+import app.simplecloud.droplet.api.auth.OAuthClient
+import app.simplecloud.droplet.api.auth.Scope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.toCollection
 import kotlinx.coroutines.reactive.asFlow
