+++ controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/AuthTokenRepository.kt
@@ -4,6 +4,7 @@ import app.simplecloud.controller.runtime.Repository
 import app.simplecloud.controller.runtime.database.Database
 import app.simplecloud.controller.shared.db.tables.records.Oauth2TokensRecord
 import app.simplecloud.controller.shared.db.tables.references.OAUTH2_TOKENS
+import app.simplecloud.droplet.api.auth.OAuthToken
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.toCollection
 import kotlinx.coroutines.reactive.asFlow
@@ -13,7 +14,7 @@ import org.jooq.exception.DataAccessException
 import java.time.Duration
 import java.time.LocalDateTime
 
-class AuthTokenRepository(private val database: Database): Repository<OAuthToken, String> {
+class AuthTokenRepository(private val database: Database) : Repository<OAuthToken, String> {
     override suspend fun getAll(): List<OAuthToken> {
         return database.context.selectFrom(OAUTH2_TOKENS)
             .asFlow()
@@ -57,14 +58,17 @@ class AuthTokenRepository(private val database: Database): Repository<OAuthToken
             element.accessToken,
             element.scope,
             element.clientId,
-            if(element.expiresIn != null) LocalDateTime.now().plusSeconds(element.expiresIn.toLong()) else null,
+            if (element.expiresIn != null) LocalDateTime.now().plusSeconds(element.expiresIn!!.toLong()) else null,
             element.userId,
         ).onDuplicateKeyUpdate()
             .set(OAUTH2_TOKENS.TOKEN_ID, element.id)
             .set(OAUTH2_TOKENS.ACCESS_TOKEN, element.accessToken)
             .set(OAUTH2_TOKENS.SCOPE, element.scope)
             .set(OAUTH2_TOKENS.CLIENT_ID, element.clientId)
-            .set(OAUTH2_TOKENS.EXPIRES_IN, if(element.expiresIn != null) LocalDateTime.now().plusSeconds(element.expiresIn.toLong()) else null)
+            .set(
+                OAUTH2_TOKENS.EXPIRES_IN,
+                if (element.expiresIn != null) LocalDateTime.now().plusSeconds(element.expiresIn!!.toLong()) else null
+            )
             .set(OAUTH2_TOKENS.USER_ID, element.userId)
             .executeAsync()
     }
