+++ controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/AuthUserRepository.kt
@@ -2,11 +2,14 @@ package app.simplecloud.controller.runtime.oauth
 
 import app.simplecloud.controller.runtime.Repository
 import app.simplecloud.controller.runtime.database.Database
-import app.simplecloud.controller.shared.auth.Scope
 import app.simplecloud.controller.shared.db.tables.records.Oauth2UsersRecord
 import app.simplecloud.controller.shared.db.tables.references.OAUTH2_TOKENS
 import app.simplecloud.controller.shared.db.tables.references.OAUTH2_USERS
 import app.simplecloud.controller.shared.db.tables.references.OAUTH2_USER_GROUPS
+import app.simplecloud.droplet.api.auth.OAuthGroup
+import app.simplecloud.droplet.api.auth.OAuthToken
+import app.simplecloud.droplet.api.auth.OAuthUser
+import app.simplecloud.droplet.api.auth.Scope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.toCollection
 import kotlinx.coroutines.reactive.asFlow
