--- controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/OAuthClient.kt
@@ -1,9 +1,0 @@
-package app.simplecloud.controller.runtime.oauth
-
-data class OAuthClient(
-    val clientId: String,
-    val clientSecret: String,
-    val redirectUri: String? = null,
-    val grantTypes: String,
-    val scope: List<String> = emptyList(),
-)
