--- controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/OAuthGroup.kt
@@ -1,6 +1,0 @@
-package app.simplecloud.controller.runtime.oauth
-
-data class OAuthGroup(
-    val scopes: List<String> = emptyList(),
-    val name: String,
-)
