--- controller-runtime/src/main/kotlin/app/simplecloud/controller/runtime/oauth/OAuthUser.kt
@@ -1,10 +1,0 @@
-package app.simplecloud.controller.runtime.oauth
-
-data class OAuthUser(
-    val groups: List<OAuthGroup> = emptyList(),
-    val scopes: List<String> = emptyList(),
-    val userId: String,
-    val username: String,
-    val hashedPassword: String,
-    val token: OAuthToken? = null,
-)
