--- controller-shared/src/main/kotlin/app/simplecloud/controller/shared/future/ListenableFutureExtension.kt
@@ -1,8 +1,0 @@
-package app.simplecloud.controller.shared.future
-
-import com.google.common.util.concurrent.ListenableFuture
-import java.util.concurrent.CompletableFuture
-
-fun <T> ListenableFuture<T>.toCompletable(): CompletableFuture<T> {
-    return ListenableFutureAdapter.toCompletable(this)
-}
