--- controller-shared/src/main/kotlin/app/simplecloud/controller/shared/secret/SecretGenerator.kt
@@ -1,15 +1,0 @@
-package app.simplecloud.controller.shared.secret
-
-import java.security.SecureRandom
-import java.util.*
-
-object SecretGenerator {
-
-    fun generate(size: Int = 64): String {
-        val random = SecureRandom()
-        val bytes = ByteArray(size)
-        random.nextBytes(bytes)
-        return Base64.getEncoder().encodeToString(bytes)
-    }
-
-}
