+++ gradle/libs.versions.toml
@@ -6,7 +6,7 @@ log4j = "2.20.0"
 protobuf = "3.25.2"
 grpc = "1.61.0"
 grpc-kotlin = "1.4.1"
-simplecloud-protospecs = "1.4.1.1.20241108142018.a431612a8826"
+droplet-api = "0.0.1-dev.66efe83"
 simplecloud-pubsub = "1.0.5"
 simplecloud-metrics = "1.0.0"
 jooq = "3.19.3"
@@ -36,7 +36,7 @@ grpc-kotlin-stub = { module = "io.grpc:grpc-kotlin-stub", version.ref = "grpc-ko
 grpc-protobuf = { module = "io.grpc:grpc-protobuf", version.ref = "grpc" }
 grpc-netty-shaded = { module = "io.grpc:grpc-netty-shaded", version.ref = "grpc" }
 
-simplecloud-protospecs = { module = "build.buf.gen:simplecloud_proto-specs_grpc_kotlin", version.ref = "simplecloud-protospecs" }
+simplecloud-droplet-api = { module = "app.simplecloud.droplet.api:droplet-api", version.ref = "droplet-api" }
 simplecloud-pubsub = { module = "app.simplecloud:simplecloud-pubsub", version.ref = "simplecloud-pubsub" }
 simplecloud-metrics = { module = "app.simplecloud:internal-metrics-api", version.ref = "simplecloud-metrics" }
 
@@ -81,7 +81,6 @@ proto = [
     "grpc-kotlin-stub",
     "grpc-protobuf",
     "grpc-netty-shaded",
-    "simplecloud-protospecs",
 ]
 jooq = [
     "jooq",
